# alem-ekub-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habte",
            "email": "habtadugna@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
...
        "/ekubs": {
            "get": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "getPageOfEkubs",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseEkubResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "updateEkub",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EkubRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "createEkub",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EkubRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ekub-users/{ekub_id}/{user_id}": {
            "put": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "removeUser",
                "parameters": [
                    {
                        "name": "ekub_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "addEkubUser",
                "parameters": [
                    {
                        "name": "ekub_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/ekub-users/{ekub-id}": {
            "post": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "joinEkub",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "unJoinEkub",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/{user-id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/payments/user/{user-id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "operationId": "getUserPayments",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/ekubs/{ekub-id}": {
            "get": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "getEkubById",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EkubResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "deleteEkub",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ekubs/search/by-name": {
            "get": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "searchEkubByName",
                "parameters": [
                    {
                        "name": "ekub-name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EkubResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/ekubs/ekub-status/{ekub-id}/{version}": {
            "get": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "getEkubStatus",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EkubStatusResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ekubs/current-round/{ekub-id}": {
            "get": {
                "tags": [
                    "ekubs"
                ],
                "operationId": "getCurrentRound",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ekub-users/member-detail/{ekub-id}/{version}": {
            "get": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "getMemberDetail",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MemberDetailResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ekub-users/equbs/{user-id}": {
            "get": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "getEkubsOfUser",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EkubResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ekub-users/ekub-users/{ekub-id}": {
            "get": {
                "tags": [
                    "ekub-users"
                ],
                "operationId": "getEkubUsers",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EkubRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    },
                    "frequencyInDays": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "penaltyPercentPerDay": {
                        "type": "number",
                        "format": "double"
                    },
                    "nextDrawDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "isExclusive": {
                        "type": "boolean"
                    },
                    "isArchived": {
                        "type": "boolean"
                    },
                    "mpesaAccountNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "telebirrAccountNumber": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "required": [
                    "amount",
                    "description",
                    "frequencyInDays",
                    "name",
                    "penaltyPercentPerDay",
                    "type"
                ]
            },
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "profilePicUrl": {
                        "type": "string"
                    },
                    "idCardImageUrl": {
                        "type": "string"
                    },
                    "guarantors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "ekubs": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ekubIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
...
            "RoundResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "ekubName": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roundNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "winner": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "paid": {
                        "type": "boolean"
                    },
                    "userGuarantees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserGuaranteeResponse"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UserGuaranteeResponse": {
                "type": "object",
                "properties": {
                    "guarantorId": {
                        "type": "string"
                    },
                    "guarantor": {
                        "type": "string"
                    },
                    "guaranteedId": {
                        "type": "string"
                    },
                    "guaranteed": {
                        "type": "string"
                    }
                }
            },
...
            "PaymentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "ekubName": {
                        "type": "string"
                    },
                    "round": {
                        "$ref": "#/components/schemas/RoundResponse"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "EkubResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "isExclusive": {
                        "type": "boolean"
                    },
                    "isArchived": {
                        "type": "boolean"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "penaltyPercentPerDay": {
                        "type": "number",
                        "format": "double"
                    },
                    "frequencyInDays": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string"
                    },
                    "roundNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMember": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "winAmount": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "nextDrawDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastDrawDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "mpesaAccountNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "telebirrAccountNumber": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "PageResponseEkubResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EkubResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "EkubStatusResponse": {
                "type": "object",
                "properties": {
                    "totalMembers": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "winnersCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}